Map the inner tests to the outer testset
`Test.jl` uses task_local_storage to store the parent test set. When
we create a distributed test, the task_local_storage will be empty and
`finish` will print the summary for the current test instead of
recording it to be used later in the summary of the outer testset.

I created a testset for each distributed test that will record the
result of the execution for the inner tests. And only call finish at
the end. Since the call to finish is the main process, the
task_local_storage isn't empty and it will record the result to be
printed later when all outer testset is finished.